Patch up multiple balances;
This is a really ugly way to get multiple balances. I should publish
get-money as it's own library on NPM and clean it up significantly with
it's own tests and have it's be it's own repo. That way it can just be a
submodule that gets included into this project and all the messy logic
like Funds / Jackie / Cash / etc can live in the index and we can treat
the lib like a lib.